+++ src/sensor_hub_service.h
@@ -30,6 +30,6 @@
     BT_UUID_16_ENCODE(0x2a6f)
     //BT_UUID_128_ENCODE(0x753e3050, 0xdf06, 0x4b53, 0xb090, 0x5e1d810c4383)
  
-void sensor_hub_update_temperature(struct bt_conn *conn, const uint8_t *data, uint16_t len);
-void sensor_hub_update_humidity(struct bt_conn *conn, const uint8_t *data, uint16_t len);
-void sensor_hub_update_pressure(struct bt_conn *conn, const uint8_t *data, uint16_t len);
+void sensor_hub_update_temperature(struct bt_conn *conn, const double *data, uint16_t len);
+void sensor_hub_update_humidity(struct bt_conn *conn, const double *data, uint16_t len);
+void sensor_hub_update_pressure(struct bt_conn *conn, const double *data, uint16_t len);
